Create globals.py and move `STRAIN_MAPPINGS_FILENAME` there
