+++ src/nplinker/globals.py
@@ -1,0 +1,1 @@
+STRAIN_MAPPINGS_FILENAME = 'strain_mappings.json'
